backend/requirements.txt
django>=4.2.0
djangorestframework>=3.14.0
Pillow>=10.0.0
numpy>=1.24.0
opencv-python>=4.8.0
face-recognition>=1.3.0 